requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.50
botocore==1.29.50
certifi==2023.5.7
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==4.1.5
django-environ==0.11.2
django-livereload-server==0.5.1
django-phonenumber-field==7.1.0
django-shortuuidfield==0.1.3
django-storages==1.13.2
env==0.1.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
phonenumberslite==8.13.15
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tornado==6.4
twilio==8.4.0
typing_extensions==4.11.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
yarl==1.9.2
